+++ security-service/security-server/src/main/java/cn/linkmore/security/service/PersonService.java
@@ -40,7 +40,7 @@ public interface PersonService {
 	 * 保存
 	 * @param person
 	 */
-	int save(ReqPerson reqPerson);
+	Long save(ReqPerson reqPerson);
 
 	/**
 	 * 更新
